+++ src/templates/sveltekit-templates.ts
@@ -1,6 +1,5 @@
 export function getClientHooksTemplate(dsn: string) {
-  return `
-import { handleErrorWithSentry, Replay } from "@sentry/sveltekit";
+  return `import { handleErrorWithSentry, Replay } from "@sentry/sveltekit";
 import * as Sentry from '@sentry/sveltekit';
 
 Sentry.init({
@@ -25,8 +24,7 @@ export const handleError = handleErrorWithSentry();
 }
 
 export function getServerHooksTemplate(dsn: string) {
-  return `
-import { sequence } from "@sveltejs/kit/hooks";
+  return `import { sequence } from "@sveltejs/kit/hooks";
 import { handleErrorWithSentry, sentryHandle } from "@sentry/sveltekit";
 import * as Sentry from '@sentry/sveltekit';
